backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
litellm
python-dotenv
loguru
